// 04-country-app/src/app/country/interfaces/rest-contries.interfaces.ts
export interface RESTCountry {
  data: Data;
}

export interface Data {
  objects: Country[];
  meta:    Meta;
}

export interface Meta {
  total:      number;
  count:      number;
  limit:      number;
  offset:     number;
  more:       boolean;
  request_id: string;
  duration:   number;
}

export interface Country {
  names:           Names;
  codes:           Codes;
  capitals:        Capital[];
  flag:            Flag;
  region:          string;
  subregion:       string;
  area:            Area;
  assets:          any[];
  borders:         string[];
  calling_codes:   string[];
  cars:            Cars;
  classification:  Classification;
  continents:      string[];
  coordinates:     Coordinates;
  currencies:      Currency[];
  date:            DateClass;
  demonyms:        Demonyms;
  economy:         Economy;
  government_type: string;
  landlocked:      boolean;
  languages:       Language[];
  leaders:         Leader[];
  links:           Links;
  memberships:     { [key: string]: boolean };
  number_format:   NumberFormat;
  parent:          Parent;
  population:      number;
  postal_code:     PostalCode;
  timezones:       string[];
  tlds:            string[];
  units:           Units;
  uuid:            string;
  _match:          Match[];
  _meta:           MetaClass;
}